### Changes since 1.5.x

#### General

- Add `Very Fast 2160p60 4K AV1`
  - *AV1 video (up to 2160p60) and AAC stereo audio, in an MP4 container.*
- Add `Very Fast 2160p60 4K HEVC`
  - *HEVC video (up to 2160p60) and AAC stereo audio, in an MP4 container.*
- Add `Fast 2160p60 4K AV1`
  - *AV1 video (up to 2160p60) and AAC stereo audio, in an MP4 container.*
- Add `Fast 2160p60 4K HEVC`
  - *HEVC video (up to 2160p60) and AAC stereo audio, in an MP4 container.*
- Add `HQ 2160p60 4K AV1 Surround`
  - *High quality AV1 video (up to 2160p60), AAC stereo audio, and Dolby Digital (AC-3) surround audio, in an MP4 container.*
- Add `HQ 2160p60 4K HEVC Surround`
  - *High quality HEVC video (up to 2160p60), AAC stereo audio, and Dolby Digital (AC-3) surround audio, in an MP4 container.*
- Add `Super HQ 2160p60 4K AV1 Surround`
  - *Super high quality AV1 video (up to 2160p60), AAC stereo audio, and Dolby Digital (AC-3) surround audio, in an MP4 container.*
- Add `Super HQ 2160p60 4K HEVC Surround`
  - *Super high quality HEVC video (up to 2160p60), AAC stereo audio, and Dolby Digital (AC-3) surround audio, in an MP4 container.*


#### Web

- Rename `Vimeo YouTube HQ` → `Creator` (all presets)
  - *High quality video for publishing via online services such as Vimeo and YouTube. H.264 video (up to `resolution`) and high bit rate AAC stereo audio in an MP4 container.*
- Remove `Vimeo YouTube 720p30`
- Rename `Gmail` → `Email 25 MB` (all presets)
  - *Up to `duration` minutes of video in 25 MB or less, for sharing via email services such as Gmail. H.264 video (up to `resolution`) and AAC stereo audio, in an MP4 container.*
- Rename `Discord Nitro Large 3-6 Minutes 1080p30` → `Social 100 MB 5 Minutes 1080p30`
  - *Up to 5 minutes of video in 100 MB or less, for sharing via online social communities such as Discord. H.264 video (up to 1080p30) and AAC stereo audio, in an MP4 container.*
  - Increase average bit rate from 2100 kb/s to 2400 kb/s, maximum bit rate from 21000 kb/s to 24000 kb/s
- Rename `Discord Nitro Medium 5-10 Minutes 720p30` → `Social 50 MB 5 Minutes 720p30`
  - *Up to 5 minutes of video in 50 MB or less, for sharing via online social communities such as Discord. H.264 video (up to 720p30) and AAC stereo audio, in an MP4 container.*
  - Can also be used to encode 10 minutes of 720p30 video in 100 MB or less. Science!
- Rename `Discord Nitro Small 10-20 Minutes 480p30` → `Social 50 MB 10 Minutes 480p30`
  - *Up to 10 minutes of video in 50 MB or less, for sharing via online social communities such as Discord. H.264 video (up to 480p30) and AAC stereo audio, in an MP4 container.*
- Rename `Discord Small 2 Minutes 360p30` → `Social 8 MB 3 Minutes 360p30`
  - *Up to 3 minutes of video in 8 MB or less, for sharing via online social communities such as Discord. H.264 video (up to 360p30) and AAC stereo audio, in an MP4 container.*
  - Change encoder speed preset from `medium` to `slow`
  - Reduce average bit rate from 450 kb/s to 270 kb/s, maximum bit rate from 4500 kb/s to 2700 kb/s
  - The idea here is to consolidate the previous Discord Small and Discord Tiny presets; bit rate and duration limit are roughly a mid point between the two
- Remove `Discord Tiny 5 Minutes 240p30`


#### Devices

- Revise description for `Amazon Fire 2160p60 4K HEVC Surround` to specify compatibility with 3rd gen devices and newer instead of 2nd gen, which according to the manufacturer only supports up to 2160p30
- Revise description for `Apple 2160p60 4K HEVC Surround` to specify iPhone 7 and later, instead of a much longer list

#### Other

- Update all HEVC presets to use 10-bit encoder and custom additional encoder options string, matching the new 4K general presets

<hr>

<details>
<summary>Original description (outdated)</summary>

New Presets:
 - Very Fast 2160p60 4K
 - Fast 2160p60 4K
 -  HQ 2160p60 4K
 -  Super HQ 2160p60 4K
    
Notes:
- The existing default preset of Fast 1080p30 remains as default since 1080p is still the dominant use case and allows for users to have to not worry about compatibility as much.

- I've also changed the audio bitrate on the 1080p Fast (192kbit), HQ (256kbit) and Super HQ presets (256kbit).  The 4K presets use the same bitrates for audio.

- Decomb and Interlace detection is turned OFF for 4K presets. The performance hit isn't work it and seemingly little 4K content is actually interlaced. 


Other Thoughts:

- Contemplating whether we should set Deinterlace detection to "Fast" for 1080p Presets out the box. It's defiantly considerably faster however I've not yet tested if it's viable with interlaced content. Will do so although if anyone has any experience on this let me know.


Open to any other ideas for preset tweaks while I'm at it.
</details>