@bradleysepos 2021-10-16
Initial thoughts:

My goal with the general presets has been maximum compatibility across players and devices, and a reasonably fast default preset since video encoding is slow by nature. So of course, my comments will be through this lens. 😸

Need to verify whether > 160 kb/s audio is widely compatible. While I suspect it's fine for anything supporting 4K, regardless of manufacturer, Apple still lists this as the max for various devices despite them supporting higher, and iTunes is a real stickler about not syncing such files sometimes.

I'm not a fan of changing this for the existing 1080p presets. That said, if we go this route, I suggest maintaining 160 kb/s for Very Fast and Fast, and going to 256 kb/s for HQ and Super HQ (though some might argue the stereo track is a fallback track and this only bloats the file size). Keeps it simple, avoids changing the default preset, and means down the road if/when we switch the default to 4K, we're only changing video stuff.

Given how much archival stuff people do in general, I suspect changes to existing behavior will cause some pushback. I don't even want to think about benchmark comparisons ("but it has the same name!").

Before we disable interlace detection and removal outside of the hardware category, I'd like to see us implement what we've discussed in the past (and again recently on IRC), some smart behavior to only run these on <= 1080p. Most recently, I suggested doing this as a set of tunes for the detection filter, such as Always Detect, HD/SD Only, and SD Only. Custom cutoff resolution could be specified manually via custom options, of course. HD/SD Only would be a great default for all existing and new HandBrake presets > 1080p.