@sr55 2021-10-21
Yeh, Lets leave the existing presets alone as it'll catch folks off guard that don't read the change note anyway. All Reverted.

I was debating the use of x265.  There was 2 reasons I kind of went against that idea.

1. The reset of the General presets are AVC and we'd probably have to add HEVC into the preset name to make that clear if we wanted to do that.
2. iTunes on Windows is a no-go for HEVC the last time I checked so keeping with the compatibility theme I figured I'd stick with AVC for these.
3. x265 is about 1/3rd the speed which didn't really meet the "very fast" speed name. 


Happy to hold off until we get the Smart Comb Detect presets in place and update this post.


I'll need to try find some more interlaced content to test with but I suspect "Fast" interlace detect might not be good enough to use in the presets. (I.e at the 1080p level and below.)