@bradleysepos 2021-10-22
> Yeh, Lets leave the existing presets alone as it'll catch folks off guard that don't read the change note anyway. All Reverted.

👍 

> I was debating the use of x265. There was 3 reasons I went against that idea.
> 
>     1. The reset of the General presets are AVC and we'd probably have to add HEVC into the preset name to make that clear if we wanted to do that.

Yes, we'd need to add HEVC to the preset name per the other presets.

>     2. iTunes on Windows is a no-go for HEVC the last time I checked so keeping with the compatibility theme I figured I'd stick with AVC for these.

That's odd, I wasn't aware of that. I was under the impression 4K AVC had compatibility issues. Will have to check my notes and external references on #1582 and prior work (linked there).

>     3. x265 is about 1/3rd the speed which didn't really meet the "very fast" speed name.

Yeah, very fast is harder to achieve there. Last I looked, you hadn't varied the encoder preset from medium, perhaps a much faster encoder preset can be tuned to produce reasonable results.

As much as I'm a fan of x264 and use it for pretty much everything, without the super transform introduced with H.265, I think the file size difference could be a bigger problem in the long run than speed.

> Happy to hold off until we get the Smart Comb Detect presets in place and update this post.

👍 

> I'll need to try find some more interlaced content to test with but I suspect "Fast" interlace detect might not be good enough to use in the presets. (I.e at the 1080p level and below.)

It's one of those things that's pretty ugly if we get it wrong, so perhaps go with the current defaults + the smart resolution bypass when implemented, and we can compare benchmarks and quality.

Thanks for your work on this! Apologies if I've come off a little *too* interested. Presets were kind of my baby for many years. Glad to have some help, especially as my time and abilities are more limited these days. Go team. 😸