@bradleysepos 2022-04-24
Rebased on current master and made some minor changes based on discussion and experimentation.

I've also been doing some more x265 tests to find a potentially better speed/size balance there. I don't think most users are going to be happy x264 yielding 25 Mbit/s video for an high quality 4K preset (likely pushing 40 Mbit/s for Super HQ). With x265 and appropriate encoder preset/settings, we can achieve ~50% performance for a 30-40% file size reduction. Tough call but I think the file size is going to be a bigger concern, especially at the higher quality end.

From my testing it's looking like for Super HQ, using roughly the same settings we use for existing 4K HEVC presets with a lower CRF value is spot on. Same encoding speed as existing 4K HEVC presets so no surprises. Ramp up the CRF and faster encoder presets for HQ, Fast, and Very Fast, so it only gets better from a performance standpoint. This would mean we won't be introducing any presets slower than ones we already have, and at the same time deliver smaller file sizes across the board and keep the higher quality presets from blowing out entirely.

Will push some additions here once I'm done testing/tweaking for others to try out.